backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.0
google-generativeai==0.3.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
supabase>=2.10.0
requests==2.31.0
python-dotenv==1.0.0
huggingface_hub>=0.20.0
redis==5.0.1
stripe>=7.0.0
razorpay>=1.4.0













